moving pages to Pages folder structure
