+++ webui/app-tracker/src/App.js
@@ -1,10 +1,10 @@
 import "../src/styles/app.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Home from "./components/Home";
-import SearchApplications from "./components/SearchApplications";
-import AddApplication from "./components/AddApplication";
+import Home from "./components/Pages/Home/Home";
+import SearchApplications from "./components/Pages/Search/SearchApplications";
+import AddApplication from "./components/Pages/AddApplication/AddApplication";
 import NotFound from "./components/NotFound";
-import Statistics from "./components/Statistics";
+import Statistics from "./components/Pages/Statistics/Statistics";
 
 function App() {
   return (
