+++ webui/app-tracker/src/components/Pages/AddApplication/AddApplication.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Menu from "./Menu/Menu";
+import Menu from "../../Menu/Menu";
 
 export default function AddApplication() {
   return (
