+++ webui/app-tracker/src/components/Pages/Search/SearchApplications.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Menu from "./Menu/Menu";
+import Menu from "../../Menu/Menu";
 
 export default function SearchApplications() {
   return (
